requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
asyncio==3.4.3
python-dotenv==1.0.0
requests==2.31.0
